pdf2text_final/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
PyMuPDF==1.24.14
gunicorn==22.0.0
pytesseract==0.3.13
Pillow==11.0.0
pdf2image==1.17.0
python-docx==1.1.2
chardet==5.2.0
fastapi==0.115.6
uvicorn==0.34.0
requests==2.32.3
python-dotenv==1.0.1
httpx==0.28.1
psutil==6.1.1

# For MySQL database (MAMP)
PyMySQL==1.1.0